packs: Download packs in parallel
Use download.c parallel download api to download packs in parallel,
the same way we do for fullfiles.

Some changes were necessary in update/slow-server test. As we are using the
parallel download api to download packs the resume  strategy is now the same
described in b07285. Because of that download resume is only enables when the
download is interrupted in the first try. So we force a download fail and a
download retry in server.py to check if swupd is trying to resume the download.
Because of this change the update/slow-server test is performing 2 retries
(first downloaded is interrupted to force a resume and the second is interrupted
because the slow-server don't support download resumes) and this adds a 30 seconds
delay to the test execution. This delay isn't going to slow down the test execution
considerably, but the ideal is in the future to support running tests with 0s
timeouts. For this reason update/slow-server is being skip when not running in
travis ci.

+++ test/functional/update/slow-server/test.bats
@@ -8,6 +8,12 @@ port=""
 targetfile="6c27df6efcd6fc401ff1bc67c970b83eef115f6473db4fb9d57e5de317eba96e"
 
 setup() {
+  # Skip this test if not running in Travis CI, because test takes too long for
+  # local development. To run this locally do: TRAVIS=true make check
+  if [ -z "${TRAVIS}" ]; then
+    skip "Skipping slow test for local development, test only runs in CI"
+  fi
+
   for i in $(seq 8081 8181); do
     pushd "$DIR"
     "./server.py" $i &
@@ -24,7 +30,6 @@ setup() {
 
   #create partial tarball
   mkdir "$DIR/state/"
-  dd if=/dev/zero of="$DIR/state/pack-os-core-from-10-to-100.tar" bs=1 count=1 >/dev/null 2>/dev/null
   chown_root "$DIR/state"
   sudo chmod 700  "$DIR/state"
 
@@ -41,8 +46,11 @@ setup() {
 }
 
 teardown() {
-  sudo rm -rf "$DIR/target-dir/usr/bin"
-  kill $server_pid
+  # teardown only if in travis CI
+  if [ -n "${TRAVIS}" ]; then
+    sudo rm -rf "$DIR/target-dir/usr/bin"
+    kill $server_pid
+  fi
 }
 
 @test "update --download with a slow server" {
